added methods to handle invalid input and clarify use

+++ DriverProgram.cpp
@@ -10,6 +10,44 @@
 #include <iomanip>
 using namespace std;
 
+void displayMenu()
+{
+	
+	cout << "~~~~~~Options~~~~~~" 			 << endl;
+	cout << "1. Find Top Scores for Event of Choice" << endl;
+	cout << "2. Delete gymnast from array"		 << endl;
+	cout << "3. Print certain gymnast's scores"	 << endl;
+	cout << "4. Find total deductions for gymnast"	 << endl;
+	cout << "5. Print team scores"		         << endl;
+	cout << "6. Print remaining gymnasts"		 << endl;
+	cout << "7. Quit"			    	 << endl;
+}
+
+
+void displayEvents()
+{	
+	cout << "~~~~~~Events~~~~~~" << endl;
+	cout << "balance beam"	     << endl;
+	cout << "floor exercise"     << endl;
+	cout << "uneven bars"  	     << endl;
+	cout << "vault"		     << endl;
+	cout << endl;
+}
+
+bool checkEventExists(string userSelection)
+{
+	string events[] = {"balance beam", "floor exercise", "uneven bars", "vault"};
+
+	for (int i = 0; i < 4; i++) {
+		
+		if (events[i] == userSelection)
+			return true;
+	}
+
+	return false;
+}
+
+
 int main(int argc, char* argv[])
 {
 	string userInput;
@@ -57,49 +95,76 @@ int main(int argc, char* argv[])
 		}
 		while(userInput != "7")
 		{
-			cout<<"~~~~~~Options~~~~~~"<<endl;
-			cout<<"1. Find Top Scores for Event of Choice"<<endl;
-			cout<<"2. Delete gymnast from array"<<endl;
-			cout<<"3. Print certain gymnast's scores"<<endl;
-			cout<<"4. Find total deductions for gymnast"<<endl;
-			cout<<"5. Print team scores"<<endl;
-			cout<<"6. Print remaining gymnasts"<<endl;
-			cout<<"7. Quit"<<endl;
+
+			cin.clear();
+
+			displayMenu();
+
 			cin>>userInput;
+			cin.ignore();
+
 			if(userInput == "1")
 			{
+
+				displayEvents();
+
 				string event;
 				cout<<"Enter event to calculate winners for:"<<endl;
-				cin.ignore();
 				getline(cin, event);
-				ScoreInstance.printTopThree(event, ScoresArray, count);
+
+				bool eventExists = checkEventExists(event);
+
+				if (eventExists)
+					ScoreInstance.printTopThree(event, ScoresArray, count);
+
+				else
+					cout << "Event does not exist! check the spelling!" << endl;
 			}
 			else if(userInput == "2")
 			{
 				int gymnastNumber;
 				cout<<"Enter gymnast number to delete:"<<endl;
 				cin>>gymnastNumber;
+
+				bool gymnastExists = ScoreInstance.checkGymnastExists(ScoresArray, gymnastNumber, count);
+
+				if (gymnastExists) {
 				//subtract one from total amount of items to compensate for gymnast deleted from array
-				count--;
-				ScoresArray = ScoreInstance.deleteGymnast(ScoresArray, gymnastNumber, count);
-				cout << "Removing gymnast number " << gymnastNumber << endl;
+					count--;
+					ScoresArray = ScoreInstance.deleteGymnast(ScoresArray, gymnastNumber, count);
+					cout << "Removing gymnast number " << gymnastNumber << endl;
+				}
 				
-				
-				
+				else
+					cout << "Gymnast does not exist!" << endl;	
 			}
 			else if(userInput == "3")
 			{
 				int gymnastNumber;
 				cout<<"Enter gymnast's number to find:"<<endl;
 				cin>>gymnastNumber;
-				ScoreInstance.findGymnast(ScoresArray, gymnastNumber, count);
+				
+				bool gymnastExists = ScoreInstance.checkGymnastExists(ScoresArray, gymnastNumber, count);
+
+				if (gymnastExists)
+					ScoreInstance.findGymnast(ScoresArray, gymnastNumber, count);
+
+				else
+					cout << "Gymnast does not exist!" << endl;
 			}
 			else if(userInput == "4")
 			{
 				int gymnastNumber;
 				cout<<"Enter gymnast to find deductions:"<<endl;
 				cin>>gymnastNumber;
-				ScoreInstance.totalDeductions(ScoresArray, count, gymnastNumber);
+	
+				bool gymnastExists = ScoreInstance.checkGymnastExists(ScoresArray, gymnastNumber, count);
+
+				if (gymnastExists)
+					ScoreInstance.totalDeductions(ScoresArray, count, gymnastNumber);
+
+				else
+					cout << "Gymnast does not exist!" << endl;
 			}
 			else if(userInput == "5")
 			{
